+++ device/api/umd/device/arc/blackhole_arc_telemetry_reader.h
@@ -5,7 +5,7 @@
  */
 #pragma once
 
-#include "umd/device/arc_telemetry_reader.h"
+#include "umd/device/arc/arc_telemetry_reader.h"
 #include "umd/device/blackhole_implementation.h"
 
 extern bool umd_use_noc1;
