+++ device/api/umd/device/arc/wormhole_arc_messenger.h
@@ -5,7 +5,7 @@
  */
 #pragma once
 
-#include "umd/device/arc_messenger.h"
+#include "umd/device/arc/arc_messenger.h"
 
 namespace tt::umd {
