+++ device/arc/arc_messenger.cpp
@@ -3,12 +3,12 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "umd/device/arc_messenger.h"
+#include "umd/device/arc/arc_messenger.h"
 
-#include "umd/device/blackhole_arc_messenger.h"
+#include "umd/device/arc/blackhole_arc_messenger.h"
+#include "umd/device/arc/wormhole_arc_messenger.h"
 #include "umd/device/tt_device/tt_device.h"
 #include "umd/device/umd_utils.h"
-#include "umd/device/wormhole_arc_messenger.h"
 
 namespace tt::umd {
