+++ device/arc/blackhole_arc_message_queue.cpp
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "umd/device/blackhole_arc_message_queue.h"
+#include "umd/device/arc/blackhole_arc_message_queue.h"
 
 #include "umd/device/tt_device/tt_device.h"
