+++ device/arc/wormhole_arc_messenger.cpp
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "umd/device/wormhole_arc_messenger.h"
+#include "umd/device/arc/wormhole_arc_messenger.h"
 
 #include <tt-logger/tt-logger.hpp>
