+++ tests/blackhole/test_arc_messages_bh.cpp
@@ -5,8 +5,8 @@
 #include <thread>
 
 #include "gtest/gtest.h"
-#include "umd/device/arc_messenger.h"
-#include "umd/device/blackhole_arc_telemetry_reader.h"
+#include "umd/device/arc/arc_messenger.h"
+#include "umd/device/arc/blackhole_arc_telemetry_reader.h"
 #include "umd/device/cluster.h"
 #include "umd/device/types/blackhole_arc.h"
