+++ tests/blackhole/test_arc_telemetry_bh.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 #include "gtest/gtest.h"
-#include "umd/device/arc_telemetry_reader.h"
+#include "umd/device/arc/arc_telemetry_reader.h"
 #include "umd/device/types/blackhole_telemetry.h"
 
 using namespace tt::umd;
