+++ tests/wormhole/test_arc_messages_wh.cpp
@@ -5,7 +5,7 @@
 #include <thread>
 
 #include "gtest/gtest.h"
-#include "umd/device/arc_messenger.h"
+#include "umd/device/arc/arc_messenger.h"
 #include "umd/device/cluster.h"
 #include "umd/device/tt_cluster_descriptor.h"
 #include "umd/device/wormhole_implementation.h"
